Moves Temperature related components to Shared - InternalTemperatureComponent and related processing has been moved to Shared - TemperatureProtectionComponent and related processing has been moved to Shared
